% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_porp.R
\name{get_porp}
\alias{get_porp}
\title{Computing points over replacement}
\usage{
get_porp(df, teams = 3L, sds = 3, plotTitle = NULL)
}
\arguments{
\item{df}{A tidy data frame or tibble containing a column "player" for the player names, a column "position" with the positions for each player, and a column "spots" with the number of roster spots available for each position. The rest of the columns should be the point projections with one column for each different projection source.}

\item{teams}{The number of teams in the draft.}

\item{sds}{The number of standard deviations away from the mean to estimate the floors and ceilings for each player.}

\item{plotTitle}{A character string providing the main title of the output plot.}
}
\value{
A list with two elements: \code{df} and \code{chart}. \code{df} is a data frame with the computed points over replacement values, and \code{chart} is a ggplot object which produces the error bar chart.
}
\description{
Computing points over replacement
}
\examples{
foo <- get_porp(bball, plotTitle = "Example Data")
head(foo$df)
foo$chart
}
