% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proj}
\alias{proj}
\title{Fantasy Basketball Point Projections}
\format{A data frame with 31 rows and 7 columns
\describe{
  \item{player}{player names}
  \item{position}{player positions: guard, forward, and center}
  \item{spots}{number of roster spots. 2 each for guards and forwards, 1 for centers}
  \item{...}{example point projections from different hypothetical sources}
}}
\usage{
proj
}
\description{
Example point projections for the January 30, 2018 DRAFT NBA main slate.
}
\keyword{datasets}
